refactor CodeState functions
