+++ packages/llmRunner/src/llmRunner/generateSummary.ts
@@ -2,8 +2,9 @@ import fs from "fs";
 import path from "path";
 
 import { Artifact, ArtifactKind } from "./Artifact";
+import { CodeState } from "./CodeState";
 import { calculatePriceMultipleCalls, LlmMetrics, LLMName } from "./LLMClient";
-import { CodeState, LLMStepInstance } from "./LLMStep";
+import { LLMStepInstance } from "./LLMStep";
 import { getLogEntryFullName, TimestampedLogEntry } from "./Logger";
 import { Workflow } from "./Workflow";
